fix: ライブ配信画面の概要部分を別コンポーネントに切り出す
