+++ web/user/src/pages/live/[...id].vue
@@ -108,17 +108,19 @@ useSeoMeta({
         <client-only>
           <the-live-video-player
             :video-src="schedule.schedule.distributionUrl"
-            :title="schedule.schedule.title"
-            :start-at="schedule.schedule.startAt"
-            :end-at="schedule.schedule.endAt"
-            :description="schedule.schedule.description"
-            :is-live-streaming="isLiveStreaming"
             :is-archive="isArchive"
-            :cordinator-id="schedule.coordinator.id"
+          />
+          <the-live-description
+            :title="schedule.schedule.title"
+            :description="schedule.schedule.description"
+            :is-archive="isArchive"
+            :is-live-streaming="isLiveStreaming"
+            :start-at="schedule.schedule.startAt"
             :marche-name="schedule.coordinator.marcheName"
-            :address="schedule.coordinator.city"
-            :cn-name="schedule.coordinator.username"
-            :cn-img-src="schedule.coordinator.thumbnailUrl"
+            :cordinator-id="schedule.coordinator.id"
+            :cordinator-name="schedule.coordinator.username"
+            :cordinator-img-src="schedule.coordinator.thumbnailUrl"
+            :cordinator-address="schedule.coordinator.city"
             @click:cordinator="handleCLickCorodinator"
           />
           <the-live-timeline
